# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Maiko API Documentation",
        "contact": {
            "email": "juancho637@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/maiko/public/api/v1"
        }
    ],
    "paths": {
...
        "/tanks/{id}": {
            "get": {
                "tags": [
                    "Tanques"
                ],
                "summary": "Muestra la información de un tanque",
                "operationId": "App\\Http\\Controllers\\Api\\Tank\\TankController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id del tanque",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Muestra la información asociada a un tanque.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Id no encontrado."
                    },
                    "401": {
                        "description": "Usuario no autorizado."
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Tanques"
                ],
                "summary": "Actualiza un tanque",
                "operationId": "App\\Http\\Controllers\\Api\\Tank\\TankController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id del tanque",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "client": {
                                        "type": "integer"
                                    },
                                    "internal_serial_number": {
                                        "type": "string"
                                    },
                                    "serial_number": {
                                        "type": "string"
                                    },
                                    "maker": {
                                        "type": "string"
                                    },
                                    "fabrication_year": {
                                        "type": "string"
                                    },
                                    "capacity": {
                                        "type": "string"
                                    },
                                    "large": {
                                        "type": "string"
                                    },
                                    "diameter": {
                                        "type": "string"
                                    },
                                    "head_thickness": {
                                        "type": "string"
                                    },
                                    "body_thickness": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Almacena un nuevo cliente asociado a una empresa.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "se debe especificar al menos un valor diferente para actualizar."
                    },
                    "401": {
                        "description": "Usuario no autorizado."
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/work_orders/{id}/inspections": {
            "post": {
                "tags": [
                    "Inspecciones"
                ],
                "summary": "Almacena una inspección",
                "operationId": "App\\Http\\Controllers\\Api\\WorkOrder\\WorkOrderInspectionController::store",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id de la orden de trabajo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "latitude",
                                    "longitude",
                                    "date"
                                ],
                                "properties": {
                                    "user": {
                                        "type": "integer"
                                    },
                                    "latitude": {
                                        "type": "integer"
                                    },
                                    "longitude": {
                                        "type": "integer"
                                    },
                                    "date": {
                                        "type": "string"
                                    },
                                    "tank": {
                                        "type": "integer"
                                    },
                                    "client": {
                                        "type": "integer"
                                    },
                                    "observation": {
                                        "type": "string"
                                    },
                                    "light_intensity": {
                                        "type": "string"
                                    },
                                    "humidity": {
                                        "type": "string"
                                    },
                                    "temperature": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Almacena una nueva inspección asociada a una orden de trabajo.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Usuario no autorizado."
                    },
                    "409": {
                        "description": "Error en validaciones de negocio."
                    },
                    "422": {
                        "description": "Entidad no procesable."
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    },
...
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}